+++ src/client/event.rs
@@ -49,6 +49,7 @@ impl HomeAssistantClient {
                 return Ok(());
             }
             "cover" => cover_event_to_entity_change(event.data),
+            "select" | "input_select" => select_event_to_entity_change(event.data),
             "sensor" | "binary_sensor" => sensor_event_to_entity_change(event.data),
             "climate" => climate_event_to_entity_change(event.data),
             "media_player" => media_player_event_to_entity_change(&self.server, event.data),